requirements.txt
# Core AI Libraries
openai>=1.12.0
google-generativeai>=0.3.2
anthropic>=0.18.0  # Optional: for Claude support in future

# Data Processing
pandas>=2.1.0
numpy>=1.24.0
openpyxl>=3.1.0  # For Excel export

# Machine Learning & Statistics
scikit-learn>=1.3.0
scipy>=1.11.0
statsmodels>=0.14.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0

# Configuration & Environment
python-dotenv>=1.0.0
pydantic>=2.5.0
pyyaml>=6.0.1

# Progress Tracking
tqdm>=4.66.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.25.0
nbformat>=5.9.0

# Utilities
python-dateutil>=2.8.2
requests>=2.31.0

# NLP for Justification Analysis (Optional)
sentence-transformers>=2.2.0  # For semantic similarity
rouge-score>=0.1.2  # For justification quality metrics
nltk>=3.8.0

# Logging & Monitoring
colorlog>=6.8.0
